types(i18n): move from index to core/i18n
